Formatting phone number input fields
This is a combination of store level javascript and HTML Generic code.  The comments in the file include a description and an example usage.  This is currently targeted at U.S. based 10-digit phone numbers.  There are multiple phone format versions already included but they could also be easily extended.
In the future I need to add comments to the code to make it more readable.  Someone could also do a better job refactoring this to make it easier to understand.